+++ examples/portokasse-example.js
@@ -1,5 +1,4 @@
-const { Internetmarke } = require('../lib/Internetmarke');
-const { PaymentMethod } = require('../lib/portokasse/Service');
+const { Internetmarke, PaymentMethod } = require('..');
 
 // To start this example directly make sure to build the project first:
 // npm i && npm run build
@@ -21,6 +20,6 @@ const { PaymentMethod } = require('../lib/portokasse/Service');
   // this will prevent you account from being charged in this test
   try {
     const amount = { value: 5, currency: 'EUR' }; //or: const amount = 500
-    await internetmarke.topUp(amount, PaymentMethod.Paypal)
+    await internetmarke.topUp(amount, PaymentMethod.PayPal)
   } catch { }
 })();