+++ test/Internetmarke.spec.ts
@@ -34,11 +34,11 @@ describe('Internetmarke', () => {
   });
 
   it('should make local 1C4A shopping cart methods available before init', () => {
-    expect(() =>
+    expect(() => {
       internetmarke.addItemToShoppingCart({ id: 1, price: 80 } as Product, {
         voucherLayout: VoucherLayout.FrankingZone
-      })
-    ).to.not.throw(InternetmarkeError);
+      });
+    }).to.not.throw(InternetmarkeError);
     expect(() => internetmarke.getItemFromShoppingCart(0)).to.not.throw(InternetmarkeError);
     expect(() => internetmarke.removeItemFromShoppingCart(0)).to.not.throw(InternetmarkeError);
     expect(() => internetmarke.getShoppingCartSummary()).to.not.throw(InternetmarkeError);
@@ -49,7 +49,7 @@ describe('Internetmarke', () => {
     expect(internetmarke.topUp(0, PaymentMethod.GiroPay)).to.eventually.be.rejectedWith(
       InternetmarkeError
     );
-    expect(internetmarke.getJournal(1)).to.eventually.be.rejectedWith(InternetmarkeError);
+    expect(internetmarke.getJournal({ days: 1 })).to.eventually.be.rejectedWith(InternetmarkeError);
 
     expect(internetmarke.initPortokasseService({} as any)).to.eventually.be.rejected;
   });